Fix CellInventory stack size data losses (#363)
Migrates the total item count and serialized stack sizes to `long`

+++ src/main/java/appeng/me/storage/BasicCellInventory.java
@@ -205,7 +205,7 @@ public class BasicCellInventory<T extends IAEStack<T>> extends AbstractCellInven
     }
 
     @Override
-    protected boolean loadCellItem(NBTTagCompound compoundTag, int stackSize) {
+    protected boolean loadCellItem(NBTTagCompound compoundTag, long stackSize) {
         // Now load the item stack
         final T t;
         try {
